:recycle: Split the strategy file
